+++ packages/auth/src/index.ts
@@ -4,7 +4,7 @@ export { Connection } from './connection/index.js'
 export { generateProof } from './invitation/index.js'
 
 export * as connection from './connection/index.js'
-export * as context from './context/index.js'
+export * as context from './team/context.js'
 export * as device from './device/index.js'
 export * as invitation from './invitation/index.js'
 export * as role from './role/index.js'
@@ -17,7 +17,7 @@ export * from './role/constants.js'
 export * from './connection/message.js'
 export * from './connection/types.js'
 export * from './connection/errors.js'
-export * from './context/types.js'
+export * from './team/context.js'
 export * from './device/types.js'
 export * from './invitation/types.js'
 export * from './role/types.js'
