+++ packages/auth/src/team/createTeam.ts
@@ -1,6 +1,6 @@
 import { createKeyset } from '@localfirst/crdx'
 import { TEAM_SCOPE } from './constants.js'
-import { type LocalContext } from 'context/index.js'
+import { type LocalContext } from 'team/context.js'
 import { Team } from 'team/Team.js'
 
 export function createTeam(teamName: string, context: LocalContext, seed?: string) {
