+++ packages/auth/src/team/index.ts
@@ -2,7 +2,7 @@ export * from './Team.js'
 export * from './constants.js'
 export * from './createTeam.js'
 export * from './decryptTeamGraph.js'
-export * from './getDeviceUserFromGraph.js'
+export * from '../connection/getDeviceUserFromGraph.js'
 export * from './isAdminOnlyAction.js'
 export * from './load.js'
 export * from './membershipResolver.js'
