+++ packages/auth/src/team/load.ts
@@ -1,6 +1,6 @@
 import { type Keyring, type KeysetWithSecrets, createKeyring } from '@localfirst/crdx'
 import { type TeamGraph } from './types.js'
-import { type LocalContext } from 'context/index.js'
+import { type LocalContext } from 'team/context.js'
 import { Team } from 'team/Team.js'
 
 export const load = (
